squash: moving things around
